build on macos as a lib for shadps4

+++ src/dynarec/arm64/arm64_printer.h
@@ -1,6 +1,6 @@
 #ifndef _ARM_PRINTER_H_
 #define _ARM_PRINTER_H_
 
-const char* arm64_print(uint32_t opcode, uint64_t addr);
+const char* arm64_print(uint32_t opcode, uintptr_t addr);
 
 #endif //_ARM_PRINTER_H_
